re-organize mrpc directory
