--- src/mrpc/src/salloc/mod.rs
@@ -1,5 +1,0 @@
-pub(crate) mod rheap;
-pub use rheap::ReadHeap;
-
-pub use salloc::backend::Error;
-pub use salloc::backend::SA_CTX;
